# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"HouseResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"size": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"hasLiked": {
						"type": "boolean"
					},
					"hasOrdered": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"imageUrl",
					"caption",
					"price",
					"address",
					"state",
					"city",
					"zipCode",
					"size",
					"createdAt",
					"userId",
					"username",
					"avatarUrl",
					"hasLiked",
					"hasOrdered"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserProfileResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"bio",
					"avatarUrl",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserOrdersResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"size": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"hasLiked": {
						"type": "boolean"
					},
					"hasOrdered": {
						"type": "boolean"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"imageUrl",
					"caption",
					"price",
					"address",
					"state",
					"city",
					"zipCode",
					"size",
					"createdAt",
					"userId",
					"username",
					"avatarUrl",
					"hasLiked",
					"hasOrdered"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Airbnb Simplified API",
		"version": "1.0.0",
		"description": "API docs for simplified Airbnb backend",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Your Name"
		}
	},
	"paths": {
...
		"/users/{userId}/profile": {
			"get": {
				"operationId": "GetUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/orders": {
			"get": {
				"operationId": "GetUserOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOrdersResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/likes": {
			"get": {
				"operationId": "GetUserLikes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/HouseResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}